[buck2]REMOVE: BuildEvent
