+++ orion-server/src/model/build_events.rs
@@ -1,38 +1,30 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.19
+use chrono::Utc;
+use serde::Serialize;
+use utoipa::ToSchema;
 
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "build_events")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub task_id: Uuid,
+#[derive(ToSchema, Serialize)]
+pub struct BuildEventDTO {
+    pub id: String,
+    pub task_id: String,
     pub retry_count: i32,
     pub exit_code: Option<i32>,
     pub log: Option<String>,
     pub log_output_file: String,
-    pub start_at: DateTimeWithTimeZone,
-    pub end_at: Option<DateTimeWithTimeZone>,
+    pub start_at: String,
+    pub end_at: Option<String>,
 }
 
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::orion_tasks::Entity",
-        from = "Column::TaskId",
-        to = "super::orion_tasks::Column::Id",
-        on_update = "Cascade",
-        on_delete = "Cascade"
-    )]
-    OrionTasks,
-}
-
-impl Related<super::orion_tasks::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::OrionTasks.def()
+impl From<&callisto::build_events::Model> for BuildEventDTO {
+    fn from(model: &callisto::build_events::Model) -> Self {
+        Self {
+            id: model.id.to_string(),
+            task_id: model.task_id.to_string(),
+            retry_count: model.retry_count,
+            exit_code: model.exit_code,
+            log: model.log.clone(),
+            log_output_file: model.log_output_file.clone(),
+            start_at: model.start_at.to_string(),
+            end_at: model.end_at.map(|dt| dt.with_timezone(&Utc).to_string()),
+        }
     }
 }
-
-impl ActiveModelBehavior for ActiveModel {}
